+++ test/node_test.dart
@@ -381,5 +381,25 @@ void main() {
         expect(scm.nominatedNodes, [node]);
       });
     });
+
+    group('saveGraphToFile', () {
+      test('should save the graph to a file', () async {
+        final node = ButterFlyExample(withScopes: true).x;
+        const path = 'test/graphs/graph_test/node_test_saveGraphToFile.dot';
+        await node.saveGraphToFile(path);
+      });
+    });
+
+    group('graph', () {
+      test(
+        'should return a dot representation of node and its suppliers '
+        'and customers',
+        () {
+          final node = ButterFlyExample(withScopes: true).x;
+          final graph = node.graph();
+          expect(graph, isNotNull);
+        },
+      );
+    });
   });
 }